Working on some refactoring of Solution precondition checks.
